+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased][unreleased] ##
+### Fixed ###
+- Removed `var_dump` in case no hits were found
+
+### Added ###
+- Exception if artist lookup yields no results
+
+
 ## 0.0.1 - 2015-05-05 ##
 ### Added ###
 - Initial client release